requirements.txt
Click==7.0
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
numpy==1.15.4
Pillow==5.3.0
